Test/requirements.txt
python-telegram-bot
python-dotenv
tinkoff-investments
aiohttp
cryptography
matplotlib
pandas
numpy
jinja2
aiohttp_jinja2